fix: redundant dom update
Here at line 49, we used SidebarItem as an alias of Sidebar.Item. As every time the render function is called, the lambda is a brand new lambda despite the code is the came. To avoid React thinking it's a different component and blow up everything in the DOM tree, we use useCallback to preserve the first copy of the lambda. However, at line 58, where we should keep the dependency list empty, one variable currentView is put, which would make the Component itself a new one when the state currentView changed, thus the whole dom tree is rebuilt.

There're two solutions here.
a) Remove the currentView dependency and make it an parameter instead. This makes the code even uglier, so I don't actually prefer to do so.

b) Make it a "component snippet" (which is what this commit has done). Usually, when we have some repeated element definition in a component, we make a function that like a component in the component, but use it like a normal function. In this way, the code would be an equivalent of the code without the alias, so we won't be bothered with the rerender issue.